[fix] #23 wikilist 상태 못 받는 상황 수정
각 위키 페이지로 넘어갈때마다 list api 호출로 해결
